Getting terrain z and overlap from worker

+++ razarion-share/src/main/java/com/btxtech/shared/dto/TerrainObjectPosition.java
@@ -63,7 +63,7 @@ public class TerrainObjectPosition {
         return this;
     }
 
-    public Matrix4 createModelMatrix(int z) {
+    public Matrix4 createModelMatrix(double z) {
         Matrix4 matrix4 = Matrix4.createTranslation(position.getX(), position.getY(), z);
         matrix4 = matrix4.multiply(Matrix4.createScale(scale, scale, scale));
         return matrix4.multiply(Matrix4.createZRotation(rotationZ));
